Login - 로그인 input 정리  Main -댓글 컴포넌트화
