+++ src/pages/minkyung/Main/Main_AsideData.js
@@ -1,0 +1,13 @@
+export const AsideData = [
+  { id: 1, dataName: '소개' },
+  { id: 2, dataName: '도움말' },
+  { id: 3, dataName: '홍보센터' },
+  { id: 4, dataName: 'API' },
+  { id: 5, dataName: '채용 정보' },
+  { id: 6, dataName: '개인정보처리방침' },
+  { id: 7, dataName: '약관' },
+  { id: 8, dataName: '위치' },
+  { id: 9, dataName: '인기계정' },
+  { id: 10, dataName: '해시태그' },
+  { id: 11, dataName: '언어' },
+];
